+++ NEWS.md
@@ -1,3 +1,19 @@
-# gadmr 0.1.0
+# gadmr 0
+
+## gadmr 0.2
+
+* **Rebuilt around package `sf` and allies instead of obsolete/gone packages**
+* **Updated GADM version used to 4.1**
+
+### gadmr 0.2.0
+
+* Rewrote `get_` functions around `sf` and `tidyverse`, not `sp` and `rgdal`/`rgeos`
+* Rewrote `get_shapefile` and `get_geopackage` and `get_map` as cases of unified function `get_gadm`
+* Rewrote `get_` function docs and rearranged args
+* Rebuilt `list_countries` using current GADM Web site
+
+## gadmr 0.1
+
+### gadmr 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.